Feature/pn 5462 bug (#464)
* PN-5462 aggiunta policy per permettere la delete sulla tabella NotificationsCostDynamoTableArn

* PN-5462 aggiunto check noticeCodeAlternative diverso da null

+++ src/main/java/it/pagopa/pn/delivery/svc/StatusService.java
@@ -71,7 +71,7 @@ public class StatusService {
                                     notificationCostEntityDao.deleteItem(NotificationCostEntity.builder()
                                             .creditorTaxIdNoticeCode(r.getPayment().getCreditorTaxId() +"##"+ r.getPayment().getNoticeCode())
                                             .build());
-                                    if(!r.getPayment().getNoticeCodeAlternative().isEmpty())
+                                    if(r.getPayment().getNoticeCodeAlternative() != null && !r.getPayment().getNoticeCodeAlternative().isEmpty())
                                         notificationCostEntityDao.deleteItem(NotificationCostEntity.builder()
                                                 .creditorTaxIdNoticeCode(r.getPayment().getCreditorTaxId() +"##"+ r.getPayment().getNoticeCodeAlternative())
                                                 .build());
